1.通知群发增加有效期设置，超时后不显示在公众号首页下方 2.通知群发增加删除功能，删除后首页下方不显示，模版点击显示信息已被删除 3.移动端支付工作人员增加发票选择，扫码支付 4.电子凭证增加提示语

+++ common/src/main/java/com/yumu/hexie/service/user/impl/MessageServiceImpl.java
@@ -129,29 +129,29 @@ public class MessageServiceImpl implements MessageService {
 		BeanUtils.copyProperties(message, notice, "id");
 		notice.setNoticeType(message.getMsgType());
 		notice.setOutsideKey(message.getId());
-		Notice n = noticeRepository.findByOutsideKey(message.getId());
-		if(n != null) {
-			notice.setId(n.getId());
-		}
+		List<Notice> ns = noticeRepository.findByOutsideKey(message.getId());
+		if(ns != null) {
+			for(Notice n : ns) {
+				notice.setId(n.getId());
+				//这里特殊处理，如果conntext有内容，则转换成链接形式存在在notice表的url字段
+				if(!ObjectUtils.isEmpty(message.getContent())) {
+					String url = messageUrl + "?oriApp=" + message.getAppid() + "#/message?messageId="+ message.getId();
+					notice.setUrl(url);
+				}
+				notice = noticeRepository.save(notice);
 
-		//这里特殊处理，如果conntext有内容，则转换成链接形式存在在notice表的url字段
+				List<NoticeSect> noticeSects = noticeSectRepository.findByNoticeId(notice.getId());
+				for (NoticeSect noticeSect : noticeSects) {
+					noticeSectRepository.delete(noticeSect);
+				}
 
-		if(!ObjectUtils.isEmpty(message.getContent())) {
-			String url = messageUrl + "?oriApp=" + message.getAppid() + "#/message?messageId="+ message.getId();
-			notice.setUrl(url);
-		}
-		notice = noticeRepository.save(notice);
-
-		List<NoticeSect> noticeSects = noticeSectRepository.findByNoticeId(notice.getId());
-		for (NoticeSect noticeSect : noticeSects) {
-			noticeSectRepository.delete(noticeSect);
-		}
-
-		for (String sectId : sectIds) {
-			NoticeSect noticeSect = new NoticeSect();
-			noticeSect.setNoticeId(notice.getId());
-			noticeSect.setSectId(Long.parseLong(sectId));
-			noticeSectRepository.save(noticeSect);
+				for (String sectId : sectIds) {
+					NoticeSect noticeSect = new NoticeSect();
+					noticeSect.setNoticeId(notice.getId());
+					noticeSect.setSectId(Long.parseLong(sectId));
+					noticeSectRepository.save(noticeSect);
+				}
+			}
 		}
 	}
